Revert "Fix #2"
This reverts commit fac2c036fecca0acd3f6fdfb182e216f6801c7b3.

+++ Assets/Game/GameController.cs
@@ -21,8 +21,8 @@ public class GameController : MonoBehaviour {
         int maxTilesY = Mathf.FloorToInt(GameSettings.cameraSize / GameSettings.gridUnit);
         int maxTilesX = maxTilesY * Screen.width / Screen.height;
 
-        int randomY = Random.Range(-maxTilesY, maxTilesY + 1);
-        int randomX = Random.Range(-maxTilesX, maxTilesX + 1);
+        float randomY = Random.Range(-maxTilesY, maxTilesY + 1);
+        float randomX = Random.Range(-maxTilesX, maxTilesX + 1);
 
         return new Vector3(randomX, randomY) * GameSettings.gridUnit;
     }
